// types/starship.type.ts
export interface Starship {
  name: string;
  model: string;
  starship_class: string;
  manufacturer: string;
  cost_in_credits: string;
  length: string;
  crew: string;
  passengers: string;
  max_atomosphering_speed: string;
  hyperdrive_rating: string;
  MGLT: string;
  cargo_capacity: string;
  consumeables: string;
  films: [string];
  pilots: [string];
  url: string;
  created: string;
  edited: string;
}